Consolidate types for the maintainer-site and rendering of it.
