+++ components/site/embedables/tiers/tiers-server.tsx
@@ -1,5 +1,5 @@
 import { getPublishedTiersForUser } from "@/app/services/tier-service";
-import { SiteMeta } from "@/lib/site/fetchers";
+import type { SiteMeta } from "@/lib/site/types";
 import Tiers from "./tiers";
 import { TiersEmbedSettingsProps } from "./tiers-embed-settings-props";
